+++ src/app/admin/admin-woodburnings/delete-woodburning/delete-woodburning.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { MatDialogRef } from '@angular/material';
-import { WoodburningStoreService } from '../../woodburning-store.service';
-import { WoodburningDetails } from '../../interfaces';
+import { WoodburningStoreService } from '../../../woodburning-store.service';
+import { WoodburningDetails } from '../../../interfaces';
 
 @Component({
     templateUrl: './delete-woodburning.component.html',
